Create BikeStore Sales Analysis
*PROJECT CASE STUDY*
The management wants to understand the conditions of the sales activities and gain insights into the various trends happening in the Bike Sales volume between 2016 and 2018. They also want to know the revenues per region, per store and per product category and per brand. the lists of sales representatives and customers are also needed. All these information are stored in different tables and for that reason, these tables are going to be joined in order to gather the necessary data from these tables for the analysis.